prediction_server/requirements.txt
# requirements.txt
functions-framework==3.*
pandas==1.5.3
numpy==1.25.2
lightgbm==3.3.5
google-cloud-storage==2.7.0
google-cloud-bigquery==3.4.0